Revert "link: add WP_LINK_FEATURE_ESTABLISHED to track when a link is PAUSED/ACTIVE"
This reverts commit 9def3f96d2a16b1aa1268011aad323eb3f94732b.

This was never a good idea, as it turns out. It was not used in practice because
it was breaking other things (see 370b692933634675213110048fcda6dff52eb52b)
and now it appears to be causing more problems in Lua object managers that don't
install because they are waiting for inactive links to become active.

Fixes: #518