ci: optimize the CI to run less when not needed
Some jobs do not make sense to be run on certain cases and we can save
CI cycles that way.

Most build rules can be run only on merge requests and custom branches.
On master, it only makes sense to build fedora_with_docs for the
purpose of deploying updated documentation on gitlab pages.
The analysis steps only make sense to run when the relevant files have
changed.